Documentation for internal methods

MetacommunityDynamics.BevertonHolt — Type

BevertonHolt{S} <: Model{Population,Biomass,S,Discrete}

The Beverton-Holt model is a discrete-time, deterministic model of population dynamics. It is commonly interpreted as a discrete-time version of the logistic model.

It is described by

$N_{t+1} =\frac{R_0 M}{N_t + M}N_t$

where $K = (R\_0 - 1)M$ is the carrying capacity.

MetacommunityDynamics.CompetitiveLotkaVolterra — Type

CompetitiveLotkaVolterra{S} <: Model{Community,Biomass,S,Continuous}

Competitive Lotka-Voterra.

MetacommunityDynamics.Coordinates — Type

Coordinates{T <: Number}

A Coordinates consists of a set of nodes with coordinates and associated environmental variables for each site.

MetacommunityDynamics.Coordinates — Method

Coordinates(layer::E; coords = [(rand(), rand()) for _ = 1:20]) where E<:EnvironmentLayer

Builds a Coordinates where the environmental variable is built from a single EnvironmentLayer, and optionally the set of coordinates can be passed as a keyword argument.

MetacommunityDynamics.Coordinates — Method

Coordinates(n::Integer)

Builds a coordinate set with n nodes in it.

MetacommunityDynamics.Coordinates — Method

Coordinates(; coords = nothing, env = nothing)

Builds a Coordinates, where both the coordinates and the environmental variables can be passed as keyword arguments. The environment should be a dictionary where (key,value) pairs are names of each environmnetal variable and vectors of that variable across each node in the Coordinates.

MetacommunityDynamics.Coordinates — Method

Coordinates(env::Dict{S,Vector})

Builds a coordinate set for a given environment matrix. The environmental matrix should be a matrix where each column is the vector of environmental variables for each node.

MetacommunityDynamics.Coordinates — Method

Coordinates(layers::Vector{E}; coords = [(rand(), rand()) for _ = 1:20]) where E<:EnvironmentLayer

Builds a coordinate set with environmental variables passed as a vector of EnvironmentLayers, and optionally coordinates passed as a keyword argument.

MetacommunityDynamics.Coordinates — Method

Coordinates(coords::Vector{T}) where T<:Tuple

Constructs a Coordinates from a set of coordinates coords, which is a vector of (x,y) pairs. Builds a random environment variable named :x.

MetacommunityDynamics.Discreteness — Type

Discreteness

MetacommunityDynamics.EnvironmentLayer — Type

EnvironmentLayer{T}

An EnvironmentalLayer stores a raster representation of a single environmental variable inside a matrix.

MetacommunityDynamics.EnvironmentLayer — Method

EnvironmentLayer(; generator = MidpointDisplacement(0.7), sz=(50,50))

Builds an EnvironmentalLayer with a NeutralLandscapes generator

MetacommunityDynamics.LogisticMap — Type

LogisticMap{S} <: Model{Population,Biomass,S,Continuous}

Logistic Map.

MetacommunityDynamics.LogisticModel — Type

LogisticModel{S} <: Model{Population,Biomass,S,Continuous}

Logistic Model.

MetacommunityDynamics.Measurement — Type

Measurement

The Measurement abstract type is a supertype for the different types of measurements a model describes, primarily [Biomass]: a continuous value representing relative amount of mass per species, [Abundance]: an integer valued count of individuals, and [Occupancy]: a binary value of whether is species is present at a location at a given time.

MetacommunityDynamics.Model — Type

Model{SC<:Scale,M<:Measurement,SP<:Spatialness,D<:Discreteness}

The abstract type from all models in MetacommunityDynamics. All Model's are parametric types that stores four parameters that describe the model: [Scale], [Measurement], [Spatialness], and [Discreteness].

MetacommunityDynamics.Parameter — Type

Parameter{T,N}

Yet-another Parameter struct.

MetacommunityDynamics.Population — Type

Population

The population scale refers to models that describe the population dynamics of a single species at a single local location.

MetacommunityDynamics.RickerModel — Type

RickerModel

MetacommunityDynamics.RickerStochasticityType — Type

RickerStochasticityType

MetacommunityDynamics.RosenzweigMacArthur — Type

struct RosenzweigMacArthur{S<:Spatialness} <: Model{Community,Biomass,S,Continuous}

Dynamics given by

$\frac{dR}{dt} = \lambda R \bigg(1 - \frac{R}{K}\bigg) - \frac{\alpha CR}{1 +\alpha \eta R}$

$\frac{dC}{dt} = \beta \frac{\alpha CR}{1 + \alpha \eta R} - \gamma C$

MetacommunityDynamics.Scale — Type

Scale

Abstract type that is a supertype for all model scales: [Population], [Metapopulation], [Community], and [Metacommunity]. A model's scale is what it was orginally defined as. For example, [TrophicLotkaVolterra] is a [Community] model, regardless of whether is has been convert to run on a spatial-graph via [spatialize].

MetacommunityDynamics.SpeciesPool — Type

SpeciesPool{T<:Number}

A SpeciesPool consists of a set of species and their corresponding traits.

MetacommunityDynamics.Trajectory — Type

Trajectory{S<:SciMLBase.AbstractTimeseriesSolution}

A trajectory is a single output for a Problem.

Base.size — Method

Base.size(coords::Coordinates)

Returns the number of nodes in a coordinate set coords.

MetacommunityDynamics._env_from_layer — Method

_env_from_layer(coords, layer::EnvironmentLayer)

Returns the values in the layer at the given coordinates, given the bounding-box os the coordinates is the extent of the layer.

MetacommunityDynamics.distance_matrix — Method

distance_matrix(coords::Coordinates; distance = Euclidean())

Returns a matrix of pairwise distances for all nodes in a Coordinates. The argument passed to distance must be of type Distance from Distances.jl.

MetacommunityDynamics.envdims — Method

envdims(coords::Coordinates)

Returns the dimensionality of the environmental variable associated with each node in a Coordinates coords.

MetacommunityDynamics.environment — Method

environment(coords::Coordinates)

Returns the dictionary of environmental variables of a Coordinates coords

MetacommunityDynamics.numsites — Method

numsites(coords::Coordinates)

Returns the number of nodes in a Coordinates coords.

MetacommunityDynamics.∂u — Method

∂u(bm::BevertonHolt, x)

Single time-step for the BevertonHolt model.